refactor(runtime): split builtins_multidim.rs into cohesive modules (§7-8)
`builtins_multidim.rs` was 1025 lines, over the 500-line module limit. Split it
into three files along its natural seams, each under 500 lines, with no behavior
change (the set of fn signatures is identical before/after):

- `builtins_multidim.rs` (232) — the module-level navigation/collection free
  functions: `multidim_index`, `multidim_delete`, `array_to_list`,
  `make_key_tuple`, `multidim_collect_leaves`, `has_multi_indices`.
- `builtins_multidim_assign.rs` (358) — `impl Interpreter` index-assignment
  lvalue ops: `builtin_index_assign_method_lvalue`(`_nested`),
  `multidim_assign_nested`, `builtin_assign_method_lvalue`.
- `builtins_multidim_subscript.rs` (440) — `impl Interpreter`
  push/append-through-accessor (`builtin_push_through_accessor`) and the
  subscript-adverb ops (`builtin_subscript_adverb`(`_error`)).

The `impl` groups don't call the module-level free fns, so no visibility change
is needed. Each file carries only the imports it uses (`ArrayKind` stays with
the free fns; `Symbol` with the subscript ops). Sixth slice of the §7-8
file-splitting sweep.
