+++ src/runtime/mod.rs
@@ -137,7 +137,9 @@ mod builtins_feed;
 mod builtins_io;
 mod builtins_lvalue;
 mod builtins_multidim;
+mod builtins_multidim_assign;
 mod builtins_multidim_ops;
+mod builtins_multidim_subscript;
 mod builtins_operators;
 pub(crate) mod builtins_reduce;
 mod builtins_string;
